+++ src/Variable.ts
@@ -1,4 +1,4 @@
-export class Label {
+export class Variable {
 	public constructor(public readonly name: string) {
 	}
 }