+++ src/index.ts
@@ -1,2 +1,3 @@
-export { Step } from './Step'
+export { builder } from './Builder'
 export { Label } from './Label'
+export { Variable } from './Variable'