+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
+## [0.0.3]  2022-09-05
+### Added
+- use Variable and Labels in "match()" step
+- use Variable in "return()" step
+
 ## [0.0.2]  2022-07-20
 ### Added
 - documentation